+++ src/game/scenes/OutGameBossSelectScene.ts
@@ -1,7 +1,7 @@
 import { Game } from '../Game';
 import { BaseOutGameScene } from './BaseOutGameScene';
 import { getBossData } from '../data';
-import { ModalUtils } from '../utils/ModalUtils';
+import { ToastUtils } from '../utils/ToastUtils';
 import type { BootstrapModal } from '../types/bootstrap';
 import { BossCardManager } from './managers/BossCardManager';
 import { DOMUpdater } from './utils/DOMUpdater';
@@ -157,7 +157,7 @@ export class OutGameBossSelectScene extends BaseOutGameScene {
                 
                 // エラーメッセージ表示
                 const errorMessage = error instanceof Error ? error.message : '不明なエラーが発生しました';
-                ModalUtils.showToast(
+                ToastUtils.showToast(
                     `ボスデータの読み込みに失敗しました: ${errorMessage}`,
                     'error'
                 );
