+++ src/game/scenes/OutGameOptionScene.ts
@@ -2,6 +2,7 @@ import { Game } from '../Game';
 import { BaseOutGameScene } from './BaseOutGameScene';
 import { PlayerSaveManager } from '../systems/PlayerSaveData';
 import { ModalUtils } from '../utils/ModalUtils';
+import { ToastUtils } from '../utils/ToastUtils';
 
 export class OutGameOptionScene extends BaseOutGameScene {
     constructor(game: Game) {
@@ -104,7 +105,7 @@ export class OutGameOptionScene extends BaseOutGameScene {
                         PlayerSaveManager.importSaveDataJson(jsonData);
                         
                         // 成功メッセージ
-                        ModalUtils.showToast('セーブデータのインポートが完了しました', 'success');
+                        ToastUtils.showToast('セーブデータのインポートが完了しました', 'success');
                         
                         // 画面更新
                         this.updateOptionScreen();
@@ -113,7 +114,7 @@ export class OutGameOptionScene extends BaseOutGameScene {
                         this.game.reboot();
                     } catch (error) {
                         console.error('Save data import failed:', error);
-                        ModalUtils.showToast('セーブデータのインポートに失敗しました', 'error');
+                        ToastUtils.showToast('セーブデータのインポートに失敗しました', 'error');
                     }
                 };
                 reader.readAsText(file);
@@ -138,11 +139,11 @@ export class OutGameOptionScene extends BaseOutGameScene {
             a.click();
             URL.revokeObjectURL(url);
             
-            ModalUtils.showToast('セーブデータをエクスポートしました', 'success');
+            ToastUtils.showToast('セーブデータをエクスポートしました', 'success');
             
         } catch (error) {
             console.error('Save data export failed:', error);
-            ModalUtils.showToast('セーブデータのエクスポートに失敗しました', 'error');
+            ToastUtils.showToast('セーブデータのエクスポートに失敗しました', 'error');
         }
     }
     
@@ -154,7 +155,7 @@ export class OutGameOptionScene extends BaseOutGameScene {
         if (confirmed) {
             try {
                 PlayerSaveManager.clearSaveData();
-                ModalUtils.showToast('セーブデータを削除しました', 'success');
+                ToastUtils.showToast('セーブデータを削除しました', 'success');
                 
                 // 画面更新
                 this.updateOptionScreen();
@@ -166,7 +167,7 @@ export class OutGameOptionScene extends BaseOutGameScene {
                 this.game.reboot();
             } catch (error) {
                 console.error('Save data clear failed:', error);
-                ModalUtils.showToast('セーブデータの削除に失敗しました', 'error');
+                ToastUtils.showToast('セーブデータの削除に失敗しました', 'error');
             }
         }
     }
@@ -176,7 +177,7 @@ export class OutGameOptionScene extends BaseOutGameScene {
      */
     private handleDebugModeToggle(enabled: boolean): void {
         localStorage.setItem('debug_mode', enabled.toString());
-        ModalUtils.showToast(`デバッグモードを${enabled ? '有効' : '無効'}にしました`, 'info');
+        ToastUtils.showToast(`デバッグモードを${enabled ? '有効' : '無効'}にしました`, 'info');
         
         // 設定反映のためページリロードを提示
         if (confirm('設定を反映するためにページをリロードしますか？')) {
