+++ src/game/scenes/managers/BattleDebugManager.ts
@@ -2,6 +2,7 @@ import { Player } from '../../entities/Player';
 import { Boss } from '../../entities/Boss';
 import { StatusEffect, StatusEffectType } from '../../systems/StatusEffect';
 import { ModalUtils } from '../../utils/ModalUtils';
+import { ToastUtils } from '../../utils/ToastUtils';
 import type { BootstrapModal } from '../../types/bootstrap';
 
 /**
@@ -128,11 +129,11 @@ export class BattleDebugManager {
                 boss.hp = Math.min(boss.hp, boss.maxHp);
             }
             
-            ModalUtils.showToast('変更が適用されました！', 'success');
+            ToastUtils.showToast('変更が適用されました！', 'success');
             
         } catch (error) {
             console.error('Error applying debug changes:', error);
-            ModalUtils.showToast('変更の適用中にエラーが発生しました', 'error');
+            ToastUtils.showToast('変更の適用中にエラーが発生しました', 'error');
         }
     }
